requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-pwa==2.0.1
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.2
